Separate BasePredictor and BaseInput
* Currently these two classes live in predictor
* Separate them out into their own classes to allow
for easier reasoning about the predictor code.
